# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SchemaShift",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/compile": {
      "post": {
        "tags": [
          "compiler"
        ],
        "summary": "Compile Endpoint",
        "description": "Run every pure compiler stage. Nothing is stored and no SQL is executed.",
        "operationId": "compile_endpoint_api_v1_compile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs": {
      "post": {
        "tags": [
          "compiler"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "compiler"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "compiler"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/export": {
      "get": {
        "tags": [
          "compiler"
        ],
        "summary": "Export Run",
        "operationId": "export_run_api_v1_runs__run_id__export_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(zip|json)$",
              "default": "zip",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/samples/{name}": {
      "get": {
        "tags": [
          "compiler"
        ],
        "summary": "Get Sample",
        "operationId": "get_sample_api_v1_samples__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompileOptions": {
        "properties": {
          "target_mongo_version": {
            "type": "string",
            "title": "Target Mongo Version",
            "default": "7.0"
          },
          "transactions_available": {
            "type": "boolean",
            "title": "Transactions Available",
            "default": true
          },
          "preserve_integer_ids": {
            "type": "boolean",
            "title": "Preserve Integer Ids",
            "default": false
          },
          "uuid_as": {
            "type": "string",
            "enum": [
              "binary",
              "string"
            ],
            "title": "Uuid As",
            "default": "binary"
          },
          "rows_per_table": {
            "type": "integer",
            "title": "Rows Per Table",
            "default": 50
          },
          "relationship_overrides": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "EMBED",
                "REFERENCE",
                "REF_ARRAY"
              ]
            },
            "type": "object",
            "title": "Relationship Overrides",
            "default": {}
          },
          "access_hints": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AccessHint"
            },
            "type": "object",
            "title": "Access Hints",
            "default": {}
          }
        },
        "type": "object",
        "title": "CompileOptions"
      },
      "CompileRequest": {
        "properties": {
          "schema_sql": {
            "type": "string",
            "maxLength": 200000,
            "title": "Schema Sql"
          },
          "queries_sql": {
            "type": "string",
            "maxLength": 200000,
            "title": "Queries Sql",
            "default": ""
          },
          "seed_sql": {
            "type": "string",
            "maxLength": 200000,
            "title": "Seed Sql",
            "default": ""
          },
          "options": {
            "$ref": "#/components/schemas/CompileOptions",
            "default": {
              "target_mongo_version": "7.0",
              "transactions_available": true,
              "preserve_integer_ids": false,
              "uuid_as": "binary",
              "rows_per_table": 50,
              "relationship_overrides": {},
              "access_hints": {}
            }
          }
        },
        "type": "object",
        "required": [
          "schema_sql"
        ],
        "title": "CompileRequest"
      },
      "CompileResponse": {
        "properties": {
          "overall_verdict": {
            "type": "string",
            "enum": [
              "SAFE",
              "CHANGED",
              "BROKEN"
            ],
            "title": "Overall Verdict"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          },
          "has_errors": {
            "type": "boolean",
            "title": "Has Errors"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "type": "array",
            "title": "Diagnostics"
          },
          "queries_line_offset": {
            "type": "integer",
            "title": "Queries Line Offset"
          },
          "ir_text": {
            "type": "string",
            "title": "Ir Text"
          },
          "ir_node_count": {
            "type": "integer",
            "title": "Ir Node Count"
          },
          "verdicts": {
            "items": {
              "$ref": "#/components/schemas/Verdict"
            },
            "type": "array",
            "title": "Verdicts"
          },
          "initial_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Initial Counts"
          },
          "changed_by_placement": {
            "items": {
              "$ref": "#/components/schemas/PlacementChange"
            },
            "type": "array",
            "title": "Changed By Placement"
          },
          "plan": {
            "items": {
              "$ref": "#/components/schemas/PlacementDecision"
            },
            "type": "array",
            "title": "Plan"
          },
          "plan_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Plan Warnings"
          },
          "graph_nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "title": "Graph Nodes"
          },
          "graph_edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "type": "array",
            "title": "Graph Edges"
          },
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled"
          },
          "generated": {
            "items": {
              "$ref": "#/components/schemas/GeneratedFile"
            },
            "type": "array",
            "title": "Generated"
          },
          "queries": {
            "items": {
              "$ref": "#/components/schemas/QueryTranslation"
            },
            "type": "array",
            "title": "Queries"
          },
          "untranslated_checks": {
            "items": {
              "$ref": "#/components/schemas/UntranslatedCheckDTO"
            },
            "type": "array",
            "title": "Untranslated Checks"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "overall_verdict",
          "counts",
          "has_errors",
          "diagnostics",
          "queries_line_offset",
          "ir_text",
          "ir_node_count",
          "verdicts",
          "initial_counts",
          "changed_by_placement",
          "plan",
          "plan_warnings",
          "graph_nodes",
          "graph_edges",
          "ai_enabled",
          "generated",
          "queries",
          "untranslated_checks",
          "warnings"
        ],
        "title": "CompileResponse"
      },
...
      "Diagnostic": {
        "properties": {
          "severity": {
            "type": "string",
            "enum": [
              "error",
              "warning",
              "info"
            ],
            "title": "Severity"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceSpan"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "severity",
          "code",
          "message"
        ],
        "title": "Diagnostic"
      },
...
      "GeneratedFile": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "mitigates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mitigates",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "path",
          "language",
          "purpose",
          "content"
        ],
        "title": "GeneratedFile"
      },
      "GraphEdge": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "child": {
            "type": "string",
            "title": "Child"
          },
          "parent": {
            "type": "string",
            "title": "Parent"
          },
          "cardinality": {
            "type": "string",
            "title": "Cardinality"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          },
          "on_delete": {
            "type": "string",
            "title": "On Delete"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "self_reference": {
            "type": "boolean",
            "title": "Self Reference",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "child",
          "parent",
          "cardinality",
          "columns",
          "on_delete",
          "decision"
        ],
        "title": "GraphEdge"
      },
      "GraphNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "columns": {
            "type": "integer",
            "title": "Columns"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "collection": {
            "type": "string",
            "title": "Collection"
          }
        },
        "type": "object",
        "required": [
          "id",
          "columns",
          "kind",
          "collection"
        ],
        "title": "GraphNode"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlacementChange": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "initial": {
            "type": "string",
            "enum": [
              "SAFE",
              "CHANGED",
              "BROKEN"
            ],
            "title": "Initial"
          },
          "final": {
            "type": "string",
            "enum": [
              "SAFE",
              "CHANGED",
              "BROKEN"
            ],
            "title": "Final"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "rule_id",
          "initial",
          "final"
        ],
        "title": "PlacementChange",
        "description": "A verdict whose status changed between the initial and final pass."
      },
      "PlacementDecision": {
        "properties": {
          "relationship_id": {
            "type": "string",
            "title": "Relationship Id"
          },
          "decision": {
            "type": "string",
            "enum": [
              "EMBED",
              "REFERENCE",
              "REF_ARRAY"
            ],
            "title": "Decision"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "default": 0.0
          },
          "top_factors": {
            "items": {
              "$ref": "#/components/schemas/Factor"
            },
            "type": "array",
            "title": "Top Factors",
            "default": []
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "overridden": {
            "type": "boolean",
            "title": "Overridden",
            "default": false
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "ai": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIOpinion"
              },
              {
                "type": "null"
              }
            ]
          },
          "agree": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agree"
          }
        },
        "type": "object",
        "required": [
          "relationship_id",
          "decision"
        ],
        "title": "PlacementDecision"
      },
...
      "RunCreated": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "RunCreated"
      },
      "RunDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "done",
              "failed"
            ],
            "title": "Status"
          },
          "verify": {
            "type": "boolean",
            "title": "Verify"
          },
          "schema_sql": {
            "type": "string",
            "title": "Schema Sql"
          },
          "queries_sql": {
            "type": "string",
            "title": "Queries Sql"
          },
          "seed_sql": {
            "type": "string",
            "title": "Seed Sql"
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "overall_verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Verdict"
          },
          "counts": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counts"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompileResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationReport"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "status",
          "verify",
          "schema_sql",
          "queries_sql",
          "seed_sql",
          "options",
          "seed",
          "overall_verdict",
          "counts",
          "duration_ms",
          "error",
          "result",
          "verification"
        ],
        "title": "RunDetail"
      },
      "RunPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "RunPage"
      },
      "RunRequest": {
        "properties": {
          "schema_sql": {
            "type": "string",
            "maxLength": 200000,
            "title": "Schema Sql"
          },
          "queries_sql": {
            "type": "string",
            "maxLength": 200000,
            "title": "Queries Sql",
            "default": ""
          },
          "seed_sql": {
            "type": "string",
            "maxLength": 200000,
            "title": "Seed Sql",
            "default": ""
          },
          "options": {
            "$ref": "#/components/schemas/CompileOptions",
            "default": {
              "target_mongo_version": "7.0",
              "transactions_available": true,
              "preserve_integer_ids": false,
              "uuid_as": "binary",
              "rows_per_table": 50,
              "relationship_overrides": {},
              "access_hints": {}
            }
          },
          "verify": {
            "type": "boolean",
            "title": "Verify",
            "default": false
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "schema_sql"
        ],
        "title": "RunRequest"
      },
...
      "SampleDetail": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "schema_sql": {
            "type": "string",
            "title": "Schema Sql"
          },
          "queries_sql": {
            "type": "string",
            "title": "Queries Sql"
          },
          "options": {
            "$ref": "#/components/schemas/CompileOptions"
          }
        },
        "type": "object",
        "required": [
          "name",
          "schema_sql",
          "queries_sql",
          "options"
        ],
        "title": "SampleDetail"
      },
...
      "UntranslatedCheckDTO": {
        "properties": {
          "table": {
            "type": "string",
            "title": "Table"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sql": {
            "type": "string",
            "title": "Sql"
          },
          "fragment": {
            "type": "string",
            "title": "Fragment"
          }
        },
        "type": "object",
        "required": [
          "table",
          "name",
          "sql",
          "fragment"
        ],
        "title": "UntranslatedCheckDTO"
      },
...
      "Verdict": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "ir_node_type": {
            "type": "string",
            "title": "Ir Node Type"
          },
          "status": {
            "type": "string",
            "enum": [
              "SAFE",
              "CHANGED",
              "BROKEN"
            ],
            "title": "Status"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "mitigation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mitigation"
          },
          "mongo_feature": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mongo Feature"
          },
          "conditions_evaluated": {
            "items": {
              "$ref": "#/components/schemas/ConditionTrace"
            },
            "type": "array",
            "title": "Conditions Evaluated",
            "default": []
          },
          "source_span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceSpan"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "rule_id",
          "ir_node_type",
          "status",
          "outcome",
          "reason"
        ],
        "title": "Verdict"
      },
...
    }
  }
}